[TAN-1735] Don't confirm admins automatically in user model
Why don't we need it anymore?

1. We needed to skip custom field validation when creating users from tenant templates
https://github.com/CitizenLabDotCo/citizenlab/pull/4657/files
The fix introduced `context: :form_submission`.
Surprisingly, `context: :form_submission` overrides `:update` and `:create`
contexts, and so callbacks/validations with `on: :create` are not called when
a record is saved like `User.new.save(context: :form_submission)`.
So, after this fix, admins stopped being confirmed automatically.

2. We discovered this bug and fixed it in another place,
in the admin/users controller when admin users are created/updated
https://github.com/CitizenLabDotCo/citizenlab/pull/5415/files

So, we don't need to do it in the model anymore.

This commit also removes `on: :create` to call set_confirmation_required
regardless of the user model context.