[CL-3870] Confirm email in SSO only if it's updated
Currently, when a user signs in with SSO, we find it not only by email,
but also by the provider UID (stored in Identity or in Verification).

If we find a user by the provider UID, and the provider doesn't return
email, we don't want to confirm the user's email.