[TAN-2254] Delete extraneous copies of map configs and layers
A bug in the project copy process caused map configs without mappable and their associated layers to be copied with every project, leading to an exponential increase in the number of map configs for some tenants. This migration removes all map configs without a mappable and their associated layers to bring the number of records back in check. Since map configs without mappables are meant to be transient, this should not result in data loss, though it may affect ongoing edits to surveys.