[TAN-2434] Remove check_publication_inconsistencies! to fix project updates
When a project changes folders, the `after_update` side effects check and update the assignees (default_assignee and idea assignees). This means that between saving the project and executing these side effects, the project can become temporarily invalid. The `check_publication_inconsistencies!` method was running during this period, blocking some project updates. Ideally, assignee updates should be part of the project update transaction, but addressing this is beyond the scope of this bug fix.