[TAN-3177] Allow moderators to create draft projects
The default assignee is now set in the `after_create` side effect instead of the `before_create`. This is because the moderator who created the project must first be added as a project moderator for the new project before they can be assigned as the default assignee. This can only be done after the project is saved because the project ID is required to make the user a project moderator.

Additionally, the `after_create` side effect is now included in the transaction to ensure the project is not created if the user cannot be promoted to project moderator.