Set default preview token for projects using `attribute`
After discussions in the CDM, we decided that the preferred approach to setting default values in models is now to use the `attribute` method. Compared to the `after_initialize` callback, this method has the advantage of setting the default value only when the model is initialized for the first time, and not when it is read from the database. And unlike the `before_validation` callback approach, it does not tie the presence of the token to the validation of the record.