[NO-TICKET] Fix preview mailers index
The list of preview mailers at `localhost:4000/rails/mailers` was broken because the newly introduced `PreviewMailer` interfered with the lazy loading of mailer previews.

Here's how `ActionMailer::Preview` determines the list of previews:
```ruby
def all
  load_previews if descendants.empty?
  descendants
end

def load_previews
  if preview_path
    Dir["#{preview_path}/**/*_preview.rb"].sort.each { |file| require_dependency file }
  end
end
```

The `load_previews` method is only triggered if `descendants` is empty. However, since `PreviewMailer` was defined in an autoloaded path and is a subclass of `ActionMailer::Preview`, it was loaded too early and short-circuited the loading of the other mailer previews.

The fix consists of moving the `PreviewMailer` class to the same folder as the other mailer previews. The downside is that now it needs to be required explicitly in the previewers that use it.