[TAN-3474] Fix bug preventing project copies
This commit fixes a bug that causes project copies to fail when the default assignee is not an admin.

The optimization that performs an authorization check on a shallow copy must set the project's `default_assignee` to `nil` (similarly to what is—implicitly—done in `LocalProjectCopyService#copy`). Otherwise, the copy will be invalid if the original default assignee is a project moderator, as the moderator doesn't have the permission to moderate the copy.