[TAN-4064] Fix project review request for folder moderators
If project reviews are enabled, the "Request approval" button is now shown to folder moderators for projects outside the folders they manage. For example, if a moderator manages one folder and one standalone project outside of that folder, they can publish any project within their folder without approval, but for the other project, project moderator rules apply.

Otherwise, the "Publish" button is shown, but is disabled.