[TAN-4519] Enable dislike reactions by default for common ground phases
Like (up) and neutral reactions are always allowed. For consistency, we could add another flag to enable or disable neutral reactions, but it's not strictly necessary, as this can be implicitly inferred by the front end based on the participation method (which is not the case for dislike (down) reactions).