[TAN-4478] Allow total to be 0 in `Jobs::Tracker`
It's okay to queue a job with a `total` (work estimate) of 0, and tracking should not support that. For instance, a job that copies ideas from one phase to another where the source phase has no ideas. While not particularly useful, this should not cause a failure.

Also, in some cases, we expect `total` to be used more as an estimate rather than an exact count. Depending on the job's type and implementation, it's possible for a job with a total of 0 to still perform work at runtime.