[TAN-4829] Store the file `size` in the `File` model
We now cache the file size directly in the model to allow the back end to sort and filter files by size without triggering requests to the storage back end (S3).

For now, the field isn’t required, as we believe this may simplify the migration of the other file models into the new `Files::File`.