[TAN-5078] Use `file_id` as the ID when serializing attachments as legacy files
This serializer has been used to refactor the legacy file controller to work with the new models while maintaining the exact same API. However, using the attachment ID as the resource identifier has caused issues and confusion on the front end. Here, we switch to using `file_id` to make the API more consistent.