[TAN-5588] Add `ContentImageService.configure_image_extraction`
This commit introduces `configure_image_extraction`, which allows image extraction to be hooked directly into `Imageable` models. It works by overriding the attribute that holds the content with images. We chose to use `prepend` instead of redefining the method in-place so it can be composed with other attribute decorators that follow the same approach. (Looking at you, `sanitize_multiloc`.)