[TAN-5588] Refactor image extraction to use before_validation callbacks
This commit updates the image extraction process to use before_validation callbacks instead of setter overrides because overrides do not work when multiloc values are modified in place.

It also reverts the Idea and CustomField models to use before_validation for sanitization, undoing earlier changes we made to make sanitization compatible with the override approach.