[TAN-5945] Simplify EmailBan model
The original goal of storing the original email hash was to make sure that the email stayed ban even if the normalization rules were changed. However, it makes it more complex to reason about the model and introduce other complexities when banning multiple variants of the same address. For the sake of simplicity, we will only store the normalized_email_hash and enforce unicity. The ban! has been reworked to be idempotent and update the reason/banner is a variant has already been banned.